Refactor component imports and remove unused code
