+++ app/(public)/guides/page.tsx
@@ -1,5 +1,11 @@
-import Greeting from "@/components/construction";
-
 export default function GuidesPage() {
-  return  <Greeting />;
+  return (
+    <div className="w-full max-w-6xl mx-auto">
+      <div className="mx-auto max-w-xl text-center">
+        <p className="text-3xl font-bold tracking-tight text-gray-900 dark:text-gray-100 sm:text-4xl">
+          Check out our guides to help you build your Home Theater
+        </p>
+      </div>
+    </div>
+  );
 }
